fixes from bot review
