+++ frontend/src/console/views/Play.vue
@@ -333,7 +333,7 @@ async function boot() {
 
   document.title = `${rom.name} | Play`;
   bezelSrc.value =
-    rom.ss_metadata?.bezel || getBezelImagePath(rom.platform_slug).value;
+    rom.ss_metadata?.bezel_path || getBezelImagePath(rom.platform_slug).value;
 
   // Configure EmulatorJS globals
   const supported = getSupportedEJSCores(rom.platform_slug);
